photos-rag-backend/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
appnope==0.1.4
Arpeggio==2.0.2
asttokens==2.4.1
bitarray==2.9.2
bitmath==1.3.3.1
bitstring==4.1.4
blessed==1.17.12
blinker==1.7.0
boto3==1.34.44
botocore==1.34.44
bpylist2==4.1.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
debugpy==1.8.0
decorator==5.1.1
distro==1.9.0
dnspython==2.5.0
executing==2.0.1
ExifRead==3.0.0
filelock==3.13.1
Flask==3.0.2
Flask-Cors==4.0.0
fsspec==2024.2.0
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.6
ipykernel==6.29.2
ipython==8.21.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
jupyter_client==8.6.0
jupyter_core==5.7.1
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==8.14.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
objexplore==1.6.3
openai==1.12.0
osxmetadata==1.3.4
osxphotos==0.67.3
packaging==23.2
pandas==2.2.0
parso==0.8.3
pathvalidate==3.2.0
pexpect==4.9.0
photoscript==0.3.1
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptpython==3.0.26
ptyprocess==0.7.0
pure-eval==0.2.2
py-applescript==1.0.3
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
PyExifTool==0.5.6
Pygments==2.17.2
pymongo==4.6.1
pyobjc-core==10.1
pyobjc-framework-AppleScriptKit==10.1
pyobjc-framework-AppleScriptObjC==10.1
pyobjc-framework-AVFoundation==10.1
pyobjc-framework-Cocoa==10.1
pyobjc-framework-CoreAudio==10.1
pyobjc-framework-CoreMedia==10.1
pyobjc-framework-CoreML==10.1
pyobjc-framework-CoreServices==10.1
pyobjc-framework-FSEvents==10.1
pyobjc-framework-Metal==10.1
pyobjc-framework-Photos==10.1
pyobjc-framework-Quartz==10.1
pyobjc-framework-ScriptingBridge==10.1
pyobjc-framework-Vision==10.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytimeparse2==1.7.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rich-theme-manager==0.11.0
s3transfer==0.10.0
safetensors==0.4.2
shortuuid==1.0.11
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
strpdatetime==0.3.0
sympy==1.12
tenacity==8.2.3
textX==4.0.1
tokenizers==0.15.2
toml==0.10.2
torch==2.2.0
torchvision==0.17.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.37.2
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.0.7
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.16.0
wurlitzer==3.0.3
xattr==1.1.0
xdg-base-dirs==6.0.1
